+++ formats/format_factory.py
@@ -1,12 +1,16 @@
 #!/usr/bin/env python
 
-from formats.fasta import Fasta
+from formats.fasta.fasta import Fasta
+
+
+def available_formats():
+    return ["fasta-dna", "fasta-aa"]
 
 
 def get_format(name):
-    if name == "fastaDna":
+    if name == "fasta-dna":
         return Fasta("DNA")
-    if name == "fastaAA":
+    if name == "fasta-aa":
         return Fasta("AA")
 
     raise Exception("FormatFactory.get_format() - Error: Unknown Format identifier.")
