+++ scripts/valifor
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import click
 import os.path
 import sys
