+++ setup.py
@@ -14,14 +14,14 @@ with open('requirements.txt') as f:
     required = f.read().splitlines()
 
 setup(
-    name='format validator',
+    name='format-validator',
     version=version,
     description='A small validator for formats in the NCBI context',
     long_description=readme,
     keywords=['validator', 'formats', 'bioinformatics'],
     author='Najia Ahmadi and Alexxander Ott',
     license=license,
-    scripts=[],  # entry point!! send to parser/cli
+    scripts=['scripts/valifor'],  # entry point! name of the script also sets name of the command in the cmd-line.
     install_requires=required,
     packages=find_packages(exclude='docs'),
     include_package_data=True
